fix bug

+++ size_measuring_pl.py
@@ -51,7 +51,7 @@ fis = interp1d(np.log10(int_models), std_smooth,'cubic')
 
 
 radcount = np.zeros(intrad.max() + 1)
-np.add.at(radcount, intrad, mask_center)
+np.add.at(radcount, np.int(intrad), mask_center)
 radmask = (radcount > 10)
 radcount[radcount == 0] = 1
